implement sactum to issue token

+++ app/Http/Controllers/V1/AlbumController.php
@@ -7,6 +7,7 @@ use App\Models\Album;
 use App\Http\Requests\StoreAlbumRequest;
 use App\Http\Requests\UpdateAlbumRequest;
 use App\Http\Resources\V1\AlbumResource;
+use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 class AlbumController extends Controller
@@ -14,9 +15,9 @@ class AlbumController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        return AlbumResource::collection(Album::paginate());
+        return AlbumResource::collection(Album::where('user_id', $request->user()->id)->paginate());
     }
 
     /**
@@ -24,7 +25,9 @@ class AlbumController extends Controller
      */
     public function store(StoreAlbumRequest $request)
     {
-        $album = Album::create($request->all());
+        $data = $request->all();
+        $data['user_id'] = $request->user()->id;
+        $album = Album::create($data);
 
         return new AlbumResource($album);
     }
@@ -32,8 +35,11 @@ class AlbumController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Album $album)
+    public function show(Request $request, Album $album)
     {
+        if ($request->user()->id != $album->user_id) {
+            return abort(403, 'Unauthorized');
+        }
         return new AlbumResource($album);
     }
 
@@ -42,6 +48,10 @@ class AlbumController extends Controller
      */
     public function update(UpdateAlbumRequest $request, Album $album)
     {
+        if ($request->user()->id != $album->user_id) {
+            return abort(403, 'Unauthorized');
+        }
+
         $album->update($request->all());
 
         return new AlbumResource($album);
@@ -50,8 +60,12 @@ class AlbumController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Album $album)
+    public function destroy(Request $request, Album $album)
     {
+        if ($request->user()->id != $album->user_id) {
+            return abort(403, 'Unauthorized');
+        }
+
         $album->delete();
 
         // return response('', 204);
