+++ app/Http/Controllers/V1/ImageManipulationController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\ResizeImageRequest;
 use App\Http\Resources\V1\ImageManipulationResource;
 use App\Models\Album;
 use App\Models\ImageManipulation;
+use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
@@ -18,13 +19,17 @@ class ImageManipulationController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        return ImageManipulationResource::collection(ImageManipulation::paginate());
+        return ImageManipulationResource::collection(ImageManipulation::where('user_id', $request->user()->id)->paginate());
     }
 
-    public function byAlbum(Album $album)
+    public function byAlbum(Request $request, Album $album)
     {
+        if ($request->user()->id != $album->user_id) {
+            return abort(403, 'Unauthorized');
+        }
+
         $where = [
             'album_id' => $album->id
         ];
@@ -44,10 +49,14 @@ class ImageManipulationController extends Controller
         $data = [
             'type' => ImageManipulation::TYPE_RESIZE,
             'data' => json_encode($all),
-            'user_id' => null,
+            'user_id' => $request->user()->id,
         ];
 
         if (isset($all['album_id'])) {
+            $album = Album::find($all['album_id']);
+            if ($request->user()->id != $album->user_id) {
+                abort(403, 'Unauthorized');
+            }
             $data['album_id'] = $all['album_id'];
         }
 
@@ -91,16 +100,23 @@ class ImageManipulationController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(ImageManipulation $image)
+    public function show(Request $request, ImageManipulation $image)
     {
+        if ($request->user()->id != $image->user_id) {
+            return abort(403, 'Unauthorized');
+        }
         return new ImageManipulationResource($image);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(ImageManipulation $image)
+    public function destroy(Request $request, ImageManipulation $image)
     {
+        if ($request->user()->id != $image->user_id) {
+            return abort(403, 'Unauthorized');
+        }
+
         $image->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
